fix(ci): resolve lint, security, and mypy CI failures
- Fix E501 line-too-long across 7 test files (break long constructors/decorators)
- Fix SIM117 nested-with in test_config.py (combine into single with)
- Fix I001 unsorted imports in test_server.py, test_integration.py
- Fix E401 multiple imports on one line in test_tools.py
- Fix mypy no-any-return in timeout.py (wrap min() with float())
- Fix security job: --select S overrides ignore list, add explicit --ignore S104

+++ src/resq_mcp/timeout.py
@@ -69,4 +69,4 @@ def get_polling_interval(attempt: int) -> float:
     """
     base = float(os.getenv("RESQ_POLLING_BASE_INTERVAL", "1.0"))
     cap = float(os.getenv("RESQ_POLLING_MAX_INTERVAL", "5.0"))
-    return min(base * (2**attempt), cap)
+    return float(min(base * (2**attempt), cap))
